pyproject.toml
[project]
name = "whatsappassistant-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "dotenv>=0.9.9",
    "fastapi>=0.124.4",
    "flake8>=7.3.0",
    "ipykernel>=7.1.0",
    "isort>=7.0.0",
    "langchain-groq>=1.1.1",
    "opik>=1.9.73",
    "orjson>=3.11.5",
    "pre-commit>=4.5.1",
    "pydantic>=2.12.5",
    "pymongo>=4.15.5",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "ruff>=0.14.9",
    "twilio>=9.8.8",
    "uvicorn>=0.38.0",
]
